Migrate zypper help to new command format
