+++ src/Zypper.h
@@ -355,7 +355,6 @@ private:
   shared_ptr<Options>  _commandOptions;
 };
 
-void print_main_help( const Zypper & zypper );
 void print_unknown_command_hint( Zypper & zypper );
 void print_command_help_hint( Zypper & zypper );
